new try

+++ exercises.c
@@ -123,7 +123,7 @@ paraéntesis balanceados. Retorna 1 si están balanceados,
 
 int parentesisBalanceados(char *cadena) {
    Stack* pila = create_stack(); 
-   for (size_t i = 0; cadena[i] != '\0'; i++) {
+   for (int i = 0; cadena[i] != '\0'; i++) {
       char c = cadena[i];
       if (c == '(' || c == '{' || c == '[') {
          char *ptrChar = (char*) malloc(sizeof(char));
